+++ Tests/CollectionViewBatchUpdatesTestCase.swift
@@ -24,8 +24,6 @@ class CollectionViewCrashTest: XCTestCase, DTCollectionViewManageable {
     
     override func setUp() {
         collectionView = UICollectionView(frame: CGRect(x: 0, y: 0, width: 100, height: 100), collectionViewLayout: UICollectionViewFlowLayout())
-        
-        manager.startManaging(withDelegate: self)
         manager.registerNibless(Cell.self)
     }